requirements.txt
apache-airflow
apache-airflow-providers-amazon
scikit-learn
pyspark
connexion[swagger-ui]
pandas
pendulum
virtualenv
boto3
python-dotenv
requests
pandas
pyarrow
selenium
webdriver-manager